api/requirements.txt
Flask
gunicorn
bcrypt
Flask-Session
python-dotenv
pyjwt==1.7.1
redis
requests

